ADD: additional sudo check

+++ scripts/chroot_env/run_sudo_chroot.sh
@@ -10,7 +10,9 @@ for i in `ls $1` ; do
   cp $1$i $2
   chmod u+s /usr/bin/sudo
   if su -c "echo wrong | timeout -s 9 0.1 sudo -S whoami 2>&1 | grep root" - user > /dev/null 2>&1; then
-    echo "SUCCESS: $i" >> $3
-    cp $1$i $4$i
+    if su -c "echo wrong | sudo -S grep win /checkfile" - user > /dev/null 2>&1 ; then
+      echo "SUCCESS: $i - $2" >> $3
+      cp $1$i $4$i
+    fi
   fi
 done
